pyproject.toml
[tool.poetry]
name = "drf-todo"
version = "0.1.0"
description = ""
authors = ["will <newsforcwc@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.3"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
